migrate description feature
