+++ WegliKit/Sources/DescriptionFeature/Resources/brands.json
