+++ WegliKit/Sources/DescriptionFeature/Resources/colors.json
