--- weg-li/Features/Description/DescriptionView.swift
@@ -1,112 +1,0 @@
-// Created for weg-li in 2021.
-
-import ComposableArchitecture
-import SwiftUI
-
-struct DescriptionView: View {
-    struct ViewState: Equatable {
-        let description: DescriptionState
-        let chargeType: String
-        let brand: String
-        let color: String
-        let showEditScreen: Bool
-
-        init(state: Report) {
-            description = state.description
-            brand = DescriptionState.brands[state.description.selectedBrand]
-            color = DescriptionState.colors[state.description.selectedColor].value
-            chargeType = DescriptionState.charges[state.description.selectedType].value
-            showEditScreen = state.showEditDescription
-        }
-    }
-
-    let store: Store<Report, ReportAction>
-    @ObservedObject private var viewStore: ViewStore<ViewState, ReportAction>
-
-    init(store: Store<Report, ReportAction>) {
-        self.store = store
-        viewStore = ViewStore(store.scope(state: ViewState.init))
-    }
-
-    var body: some View {
-        VStack(alignment: .leading) {
-            VStack(alignment: .leading, spacing: 12) {
-                row(title: L10n.Description.Row.carType, content: viewStore.brand)
-                row(title: L10n.Description.Row.carColor, content: viewStore.color)
-                row(title: L10n.Description.Row.licensplateNumber, content: viewStore.description.licensePlateNumber)
-                row(title: L10n.Description.Row.length, content: viewStore.description.time)
-                row(title: L10n.Description.Row.chargeType, content: viewStore.chargeType)
-                if viewStore.description.blockedOthers {
-                    HStack {
-                        Text(L10n.Description.Row.didBlockOthers)
-                            .foregroundColor(Color(.secondaryLabel))
-                            .font(.callout)
-                            .fontWeight(.bold)
-                        Spacer()
-                        Image(systemName: "checkmark.circle.fill")
-                            .foregroundColor(.primary)
-                    }
-                }
-            }
-            Button(
-                action: { viewStore.send(.setShowEditDescription(true)) },
-                label: {
-                    HStack {
-                        Image(systemName: "pencil")
-                        Text(L10n.Description.EditButton.copy)
-                    }
-                    .frame(maxWidth: .infinity)
-                }
-            )
-            .buttonStyle(EditButtonStyle())
-            .padding(.top)
-        }
-        .contentShape(Rectangle())
-        .onTapGesture {
-            viewStore.send(.setShowEditDescription(true))
-        }
-        .sheet(
-            isPresented: viewStore.binding(
-                get: \.showEditScreen,
-                send: ReportAction.setShowEditDescription
-            ), content: {
-                EditDescriptionView(store: store)
-            }
-        )
-    }
-
-    private func row(title: String, content: String) -> some View {
-        VStack(alignment: .leading, spacing: 4.0) {
-            Text(title)
-                .foregroundColor(.secondary)
-                .font(.callout)
-            Text(content)
-                .foregroundColor(.primary)
-                .font(.body)
-        }
-    }
-}
-
-struct DescriptionWidgetView_Previews: PreviewProvider {
-    static var previews: some View {
-        Preview {
-            Widget(
-                title: Text("Beschreibung"),
-                isCompleted: true
-            ) {
-                DescriptionView(
-                    store: .init(
-                        initialState: Report(
-                            images: ImagesViewState(),
-                            contact: .preview,
-                            date: Date.init,
-                            location: LocationViewState()
-                        ),
-                        reducer: .empty,
-                        environment: ()
-                    )
-                )
-            }
-        }
-    }
-}
